refactor: 流程节点系统重设计 + 代码审查修复 + caller 追踪
流程节点系统：
- 移除 start 节点，用 sequence 替代
- nodes 从数组改为 map 格式，消除自定义反序列化
- loop.body 改为单节点引用，多步骤用 sequence 包装
- 新增 break/continue 节点类型，支持循环控制
- 新增 defaultDelayMs 配置，节点级 delayMs 可覆盖
- 新增 loop.condition/breakCondition 条件循环
- weighted 使用独立 options 字段，sequence.next 改为 []string
- 移除并行执行，全部串行
- ActionHandler 接口移除 OnNodeError，日志在 executor 内输出
- Executor 新增 caller 字段用于日志追踪

代码审查修复：
- LuaAdapter.acquire() 增加 30 秒超时，防止死锁
- NewLuaAdapter 初始化失败时清理已创建的 LState
- state/store.go toInt 添加 float32 分支
- Lua 脚本错误时正确返回非 0（listen_start_loading/match_succeed/battle_end）
- 移除 api_network.go 过时的向后兼容注释
- 移除 api_utils.go 已废弃的 log_info/log_error 注册
- robot.go import 按 标准库/第三方/本地包 分组排序
- sync_frame_data/load_progress 增加 send 返回值检查

配置变更：
- flow.json 按新格式重写，覆盖完整压测流程
- config.json mainService 替代 defaultListenServer
- udpServices 默认值精简为 ["udp"]
- 删除已迁移为声明式的 20 个 Lua 脚本

+++ adapter/adapter.go
@@ -35,10 +35,10 @@ type Adapter interface {
 	// Decode 将完整数据包解码为路由键、消息体和协议头错误码。
 	// responseKey 是字符串路由键，用于请求-响应匹配和监听分发。
 	// 格式由适配器决定，典型格式："{cmd}:{act}"，如 "3:1"。
-	// headerErr 为协议头中的错误码字段（无该字段的协议返回 0）。
+	// headerErr 为协议头中的错误码，Lua decode() 必须返回数字，Go 用 uint64 接收。
 	// 非零时 Connection.OnReceive 记录告警，仍继续路由（让请求正常完成）。
 	// TCP 和 UDP 使用同一 Decode（接收侧无偏移问题）。
-	Decode(data []byte, secretKey []byte) (responseKey string, body []byte, headerErr uint16)
+	Decode(data []byte, secretKey []byte) (responseKey string, body []byte, headerErr uint64)
 
 	// ExpectedResponseKey 从发送路由计算期望的响应路由键。
 	// 用于 TCPRequest 等待响应时注册临时通道。
